Add compat decorator for V2.3
