+++ src/etc/di.xml
@@ -55,6 +55,16 @@
         </arguments>
     </type>
 
+    <type name="Emico\TweakwiseExport\Model\Write\Products\CollectionDecorator\StockData">
+        <arguments>
+            <argument name="metaData" xsi:type="object">Magento\Framework\App\ProductMetadataInterface</argument>
+            <argument name="stockDecorators" xsi:type="array">
+                <item name="V22X" xsi:type="object">Emico\TweakwiseExport\Model\Write\Products\CollectionDecorator\StockData\V22X</item>
+                <item name="V23X" xsi:type="object">Emico\TweakwiseExport\Model\Write\Products\CollectionDecorator\StockData\V23X</item>
+            </argument>
+        </arguments>
+    </type>
+
     <type name="Emico\TweakwiseExport\Model\Write\Products\Iterator">
         <arguments>
             <argument name="helper" xsi:type="object">Emico\TweakwiseExport\Model\Helper</argument>
